Cloody lizens aus TikiPak, ContentManager - LoadData für TikiPak, SpriteBatch neu geschrieben mit funktionierender LayerDepth, TikiMesh mit LocalMatrix für die Türme, Türme mit drehendem geschütz, RTS Camera verbessert mit F5 kann zur alten gewechselt werden.
git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C1125

+++ src/Sound/SoundModule.cpp
@@ -160,7 +160,7 @@ namespace TikiEngine
 			FMOD_CREATESOUNDEXINFO info;
 			ZeroMemory(&info, sizeof(info));
 			info.cbsize = sizeof(info);
-			info.length = stream->GetLength();
+			info.length = (UInt32)stream->GetLength();
 
 			system->createStream(data, FMOD_OPENMEMORY, &info, &sound);
 			//system->createStream(fileName, FMOD_DEFAULT, 0, &sound);
